# packages/contracts/openapi/control-plane.v1.yaml
openapi: 3.1.0
info:
  title: Private Cloud Control Plane API
  version: 1.0.0
  summary: Provider-neutral virtual-machine lifecycle control plane
  description: |
    Accepts durable asynchronous lifecycle commands and exposes project-scoped read models.
    A `202` response means intent and its operation record committed; it does not mean provider work completed.
  license:
    name: Apache-2.0
    identifier: Apache-2.0
servers:
  - url: https://cloud.example.invalid
    description: Environment-specific Gateway API endpoint
security:
  - bearerAuth: []
...
paths:
...
  /v1/projects/{projectId}/instances:
    parameters:
      - $ref: './components.v1.yaml#/components/parameters/ProjectId'
    post:
      operationId: createInstance
      tags: [Instances]
      summary: Accept an instance-create command
      x-audience: tenant
      x-idempotency: required
      x-max-body-bytes: 65536
      x-error-codes:
        [
          VALIDATION_FAILED,
          PROJECT_ACCESS_DENIED,
          QUOTA_EXCEEDED,
          IDEMPOTENCY_CONFLICT,
          PROFILE_DISABLED,
        ]
      x-requirements:
        [
          SRS-FR-012,
          SRS-FR-014,
          SRS-FR-015,
          SRS-FR-016,
          SRS-FR-017,
          SRS-FR-018,
          SRS-FR-019,
          SRS-FR-020,
          SRS-FR-024,
          SRS-FR-025,
        ]
      parameters:
        - $ref: './components.v1.yaml#/components/parameters/IdempotencyKey'
        - $ref: './components.v1.yaml#/components/parameters/CorrelationId'
        - $ref: './components.v1.yaml#/components/parameters/Traceparent'
        - $ref: './components.v1.yaml#/components/parameters/Tracestate'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './components.v1.yaml#/components/schemas/CreateInstanceRequest'
      responses:
        '202':
          $ref: './components.v1.yaml#/components/responses/MutationAccepted'
        '400': { $ref: './components.v1.yaml#/components/responses/BadRequest' }
        '401': { $ref: './components.v1.yaml#/components/responses/Unauthorized' }
        '403': { $ref: './components.v1.yaml#/components/responses/Forbidden' }
        '409': { $ref: './components.v1.yaml#/components/responses/Conflict' }
        '413': { $ref: './components.v1.yaml#/components/responses/PayloadTooLarge' }
        '422': { $ref: './components.v1.yaml#/components/responses/Unprocessable' }
    get:
      operationId: listInstances
      tags: [Instances]
      summary: List project instances
      x-audience: tenant
      x-idempotency: read-only
      x-error-codes: [VALIDATION_FAILED, PROJECT_ACCESS_DENIED]
      x-requirements: [SRS-FR-002, SRS-FR-037, SRS-FR-038]
      parameters:
        - $ref: './components.v1.yaml#/components/parameters/Cursor'
        - $ref: './components.v1.yaml#/components/parameters/Limit'
        - name: lifecycleState
          in: query
          schema:
            $ref: './components.v1.yaml#/components/schemas/InstanceLifecycleState'
      responses:
        '200':
          description: Instance page
          content:
            application/json:
              schema:
                $ref: './components.v1.yaml#/components/schemas/InstancePage'
        '400': { $ref: './components.v1.yaml#/components/responses/BadRequest' }
        '401': { $ref: './components.v1.yaml#/components/responses/Unauthorized' }
        '403': { $ref: './components.v1.yaml#/components/responses/Forbidden' }
  /v1/projects/{projectId}/instances/{instanceId}:
    parameters:
      - $ref: './components.v1.yaml#/components/parameters/ProjectId'
      - $ref: './components.v1.yaml#/components/parameters/InstanceId'
    get:
      operationId: getInstance
      tags: [Instances]
      summary: Read one project instance
      x-audience: tenant
      x-idempotency: read-only
      x-error-codes: [PROJECT_ACCESS_DENIED, INSTANCE_NOT_FOUND]
      x-requirements: [SRS-FR-002, SRS-FR-037, SRS-FR-038]
      responses:
        '200':
          description: Instance detail
          content:
            application/json:
              schema:
                $ref: './components.v1.yaml#/components/schemas/Instance'
        '401': { $ref: './components.v1.yaml#/components/responses/Unauthorized' }
        '403': { $ref: './components.v1.yaml#/components/responses/Forbidden' }
        '404': { $ref: './components.v1.yaml#/components/responses/NotFound' }
    delete:
      operationId: retainInstance
      tags: [Instances]
      summary: Accept a soft-delete retention command
      description: Normal deletion retains the provider resource; only an administrative purge destroys it.
      x-audience: tenant
      x-idempotency: required
      x-error-codes:
        [PROJECT_ACCESS_DENIED, INSTANCE_NOT_FOUND, INSTANCE_BUSY, IDEMPOTENCY_CONFLICT]
      x-requirements: [SRS-FR-014, SRS-FR-044, SRS-FR-045, SRS-FR-051, SRS-FR-052]
      parameters:
        - $ref: './components.v1.yaml#/components/parameters/IdempotencyKey'
        - $ref: './components.v1.yaml#/components/parameters/CorrelationId'
        - $ref: './components.v1.yaml#/components/parameters/Traceparent'
        - $ref: './components.v1.yaml#/components/parameters/Tracestate'
      responses:
        '202': { $ref: './components.v1.yaml#/components/responses/MutationAccepted' }
        '401': { $ref: './components.v1.yaml#/components/responses/Unauthorized' }
        '403': { $ref: './components.v1.yaml#/components/responses/Forbidden' }
        '404': { $ref: './components.v1.yaml#/components/responses/NotFound' }
        '409': { $ref: './components.v1.yaml#/components/responses/Conflict' }
  /v1/projects/{projectId}/instances/{instanceId}/actions:
    parameters:
      - $ref: './components.v1.yaml#/components/parameters/ProjectId'
      - $ref: './components.v1.yaml#/components/parameters/InstanceId'
    post:
      operationId: mutateInstance
      tags: [Instances]
      summary: Accept a power or resize command
      x-audience: tenant
      x-idempotency: required
      x-max-body-bytes: 65536
      x-error-codes:
        [
          VALIDATION_FAILED,
          PROJECT_ACCESS_DENIED,
          INSTANCE_NOT_FOUND,
          INSTANCE_BUSY,
          QUOTA_EXCEEDED,
          DISK_SHRINK_FORBIDDEN,
          IDEMPOTENCY_CONFLICT,
        ]
      x-requirements:
        [
          SRS-FR-014,
          SRS-FR-039,
          SRS-FR-040,
          SRS-FR-041,
          SRS-FR-042,
          SRS-FR-043,
          SRS-FR-044,
          SRS-FR-045,
        ]
      parameters:
        - $ref: './components.v1.yaml#/components/parameters/IdempotencyKey'
        - $ref: './components.v1.yaml#/components/parameters/CorrelationId'
        - $ref: './components.v1.yaml#/components/parameters/Traceparent'
        - $ref: './components.v1.yaml#/components/parameters/Tracestate'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './components.v1.yaml#/components/schemas/InstanceAction'
      responses:
        '202': { $ref: './components.v1.yaml#/components/responses/MutationAccepted' }
        '400': { $ref: './components.v1.yaml#/components/responses/BadRequest' }
        '401': { $ref: './components.v1.yaml#/components/responses/Unauthorized' }
        '403': { $ref: './components.v1.yaml#/components/responses/Forbidden' }
        '404': { $ref: './components.v1.yaml#/components/responses/NotFound' }
        '409': { $ref: './components.v1.yaml#/components/responses/Conflict' }
        '413': { $ref: './components.v1.yaml#/components/responses/PayloadTooLarge' }
        '422': { $ref: './components.v1.yaml#/components/responses/Unprocessable' }
  /v1/projects/{projectId}/instances/{instanceId}/snapshots:
    parameters:
      - $ref: './components.v1.yaml#/components/parameters/ProjectId'
      - $ref: './components.v1.yaml#/components/parameters/InstanceId'
    get:
      operationId: listSnapshots
      tags: [Snapshots]
      summary: List instance-owned snapshots
      x-audience: tenant
      x-idempotency: read-only
      x-error-codes: [VALIDATION_FAILED, PROJECT_ACCESS_DENIED, INSTANCE_NOT_FOUND]
      x-requirements: [SRS-FR-046, SRS-FR-050]
      responses:
        '200':
          description: Snapshot page
          content:
            application/json:
              schema:
                $ref: './components.v1.yaml#/components/schemas/SnapshotPage'
        '400': { $ref: './components.v1.yaml#/components/responses/BadRequest' }
        '401': { $ref: './components.v1.yaml#/components/responses/Unauthorized' }
        '403': { $ref: './components.v1.yaml#/components/responses/Forbidden' }
        '404': { $ref: './components.v1.yaml#/components/responses/NotFound' }
    post:
      operationId: createSnapshot
      tags: [Snapshots]
      summary: Accept a snapshot-create command
      x-audience: tenant
      x-idempotency: required
      x-max-body-bytes: 65536
      x-error-codes:
        [
          PROJECT_ACCESS_DENIED,
          INSTANCE_NOT_FOUND,
          INSTANCE_BUSY,
          QUOTA_EXCEEDED,
          IDEMPOTENCY_CONFLICT,
        ]
      x-requirements: [SRS-FR-014, SRS-FR-044, SRS-FR-047, SRS-FR-050]
      parameters:
        - $ref: './components.v1.yaml#/components/parameters/IdempotencyKey'
        - $ref: './components.v1.yaml#/components/parameters/CorrelationId'
        - $ref: './components.v1.yaml#/components/parameters/Traceparent'
        - $ref: './components.v1.yaml#/components/parameters/Tracestate'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './components.v1.yaml#/components/schemas/CreateSnapshotRequest'
      responses:
        '202': { $ref: './components.v1.yaml#/components/responses/MutationAccepted' }
        '400': { $ref: './components.v1.yaml#/components/responses/BadRequest' }
        '401': { $ref: './components.v1.yaml#/components/responses/Unauthorized' }
        '403': { $ref: './components.v1.yaml#/components/responses/Forbidden' }
        '404': { $ref: './components.v1.yaml#/components/responses/NotFound' }
        '409': { $ref: './components.v1.yaml#/components/responses/Conflict' }
        '422': { $ref: './components.v1.yaml#/components/responses/Unprocessable' }
  /v1/projects/{projectId}/instances/{instanceId}/snapshots/{snapshotId}/actions:
    parameters:
      - $ref: './components.v1.yaml#/components/parameters/ProjectId'
      - $ref: './components.v1.yaml#/components/parameters/InstanceId'
      - $ref: './components.v1.yaml#/components/parameters/SnapshotId'
    post:
      operationId: rollbackSnapshot
      tags: [Snapshots]
      summary: Accept a snapshot rollback command
      x-audience: tenant
      x-idempotency: required
      x-max-body-bytes: 65536
      x-error-codes:
        [
          PROJECT_ACCESS_DENIED,
          INSTANCE_NOT_FOUND,
          SNAPSHOT_NOT_FOUND,
          SNAPSHOT_OWNERSHIP_MISMATCH,
          INSTANCE_BUSY,
          IDEMPOTENCY_CONFLICT,
        ]
      x-requirements: [SRS-FR-014, SRS-FR-044, SRS-FR-048, SRS-FR-050]
      parameters:
        - $ref: './components.v1.yaml#/components/parameters/IdempotencyKey'
        - $ref: './components.v1.yaml#/components/parameters/CorrelationId'
        - $ref: './components.v1.yaml#/components/parameters/Traceparent'
        - $ref: './components.v1.yaml#/components/parameters/Tracestate'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './components.v1.yaml#/components/schemas/SnapshotRollbackRequest'
      responses:
        '202': { $ref: './components.v1.yaml#/components/responses/MutationAccepted' }
        '400': { $ref: './components.v1.yaml#/components/responses/BadRequest' }
        '401': { $ref: './components.v1.yaml#/components/responses/Unauthorized' }
        '403': { $ref: './components.v1.yaml#/components/responses/Forbidden' }
        '404': { $ref: './components.v1.yaml#/components/responses/NotFound' }
        '409': { $ref: './components.v1.yaml#/components/responses/Conflict' }
  /v1/projects/{projectId}/instances/{instanceId}/snapshots/{snapshotId}:
    parameters:
      - $ref: './components.v1.yaml#/components/parameters/ProjectId'
      - $ref: './components.v1.yaml#/components/parameters/InstanceId'
      - $ref: './components.v1.yaml#/components/parameters/SnapshotId'
    delete:
      operationId: deleteSnapshot
      tags: [Snapshots]
      summary: Accept a snapshot-delete command
      x-audience: tenant
      x-idempotency: required
      x-error-codes:
        [
          PROJECT_ACCESS_DENIED,
          INSTANCE_NOT_FOUND,
          SNAPSHOT_NOT_FOUND,
          SNAPSHOT_OWNERSHIP_MISMATCH,
          INSTANCE_BUSY,
          IDEMPOTENCY_CONFLICT,
        ]
      x-requirements: [SRS-FR-014, SRS-FR-044, SRS-FR-049, SRS-FR-050]
      parameters:
        - $ref: './components.v1.yaml#/components/parameters/IdempotencyKey'
        - $ref: './components.v1.yaml#/components/parameters/CorrelationId'
        - $ref: './components.v1.yaml#/components/parameters/Traceparent'
        - $ref: './components.v1.yaml#/components/parameters/Tracestate'
      responses:
        '202': { $ref: './components.v1.yaml#/components/responses/MutationAccepted' }
        '401': { $ref: './components.v1.yaml#/components/responses/Unauthorized' }
        '403': { $ref: './components.v1.yaml#/components/responses/Forbidden' }
        '404': { $ref: './components.v1.yaml#/components/responses/NotFound' }
        '409': { $ref: './components.v1.yaml#/components/responses/Conflict' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT